+++ documents/scripts/tables/contacts.sql
@@ -7,14 +7,14 @@
 -- =====================================================
 
 CREATE TABLE IF NOT EXISTS public.contacts (
-    id SERIAL PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     platform VARCHAR(255) NOT NULL,
     url VARCHAR(512),
     description TEXT,
     banner VARCHAR(512),
     created_at TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
     updated_at TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
-    profile_id SERIAL NOT NULL,
+    profile_id BIGINT NOT NULL,
     CONSTRAINT fk_contacts_profile FOREIGN KEY (profile_id) REFERENCES profiles(id) ON DELETE CASCADE
 );
